requirements.txt
bingads==13.0.14 ; python_version >= "3.9" and python_version < "3.10"
cachetools==5.2.0 ; python_version >= "3.9" and python_version < "3.10"
certifi==2022.9.24 ; python_version >= "3.9" and python_version < "3.10"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "3.10"
compose==1.4.8 ; python_version >= "3.9" and python_version < "3.10"
future==0.18.2 ; python_version >= "3.9" and python_version < "3.10"
google-api-core==2.10.1 ; python_version >= "3.9" and python_version < "3.10"
google-api-core[grpc]==2.10.1 ; python_version >= "3.9" and python_version < "3.10"
google-auth==2.12.0 ; python_version >= "3.9" and python_version < "3.10"
google-cloud-bigquery==2.34.4 ; python_version >= "3.9" and python_version < "3.10"
google-cloud-core==2.3.2 ; python_version >= "3.9" and python_version < "3.10"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "3.10"
google-resumable-media==2.3.3 ; python_version >= "3.9" and python_version < "3.10"
googleapis-common-protos==1.56.4 ; python_version >= "3.9" and python_version < "3.10"
grpcio-status==1.48.2 ; python_version >= "3.9" and python_version < "3.10"
grpcio==1.49.1 ; python_version >= "3.9" and python_version < "3.10"
idna==3.4 ; python_version >= "3.9" and python_version < "3.10"
packaging==21.3 ; python_version >= "3.9" and python_version < "3.10"
proto-plus==1.22.1 ; python_version >= "3.9" and python_version < "3.10"
protobuf==3.20.2 ; python_version >= "3.9" and python_version < "3.10"
pyasn1-modules==0.2.8 ; python_version >= "3.9" and python_version < "3.10"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "3.10"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "3.10"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.10"
requests==2.28.1 ; python_version >= "3.9" and python_version < "3.10"
rsa==4.9 ; python_version >= "3.9" and python_version < "3.10"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.10"
suds-community==1.1.2 ; python_version >= "3.9" and python_version < "3.10"
urllib3==1.26.12 ; python_version >= "3.9" and python_version < "3.10"
